Refactor: Removed the Networking file
Instead of using a Networking file, refactored to be using the Old API file
